Master tests (#149)
* fixed generate meta tests

* fixed inference script tests

* fixed metrics tests

* fixed preprocessing (masks, tile, mp aux)

* fixed evaluation metrics tests

* lots of issues with inference. fixed 2d so far

* fixed 2.5d inference, still working on 3d

* 3 remaining errors in 3d inference

* debugging 3d inference

* debugging 3d inference

* debugging 3d inference, fixed overlap shape

* fixed 5d tiling bug

* fixed 3d inference!

* fixed stitch tests

* fixed dataset tests

* fixed dataset w mask tests

* fixed inference dataset tests

* fixed plot utils tests

* added flatfield tests

* fixed gen mask tests

* working on tiling tests

* fixed tile nonuni tests

* fixed uniform tile tests

* fixed aux utils tests

* debugging image utils

* updated pandas version to avoid attribute error in pandas

* fixed image utils tests

* fixed mask utils tests

* fixed mp utils tests

* debugging tile utils

* fixed tile utils tests

* newer mpl version

* updated skimage, debugging flatfield

* sort output of os.listdir

* Inference features (progressbar, define prediction folder name, metrics in figures) (#147)

* minor fix for preprocessing_script

* changed the layout of sub panels in predicted figures

* removed redundant slice margin adjustment

* disable the check for modelcheckpoint monitor

* Added flags to save predicted images in image directory or model directory

* Added data normalization options to preprocessing

* updated image_inference script with data normalization

* bug fix

* updated tests

* updated tests and 3d preprocessing

* added dataset otsu mask type

* added script for pooling multiple datasets

* created meta_utils; added multi-processing option to meta_generator; estimate dataset z-score parameters from foreground images only

* added function to sample values at block centers

* added blocks_meta.csv

* bug fix

* updated functions to sample pixels and compute zscore parameters

* Fixed the bug caused by mixed numpy datatypes; make data normalization backward compatible

* updated metrics_script to new normalization options
added get_pp_config function to preprocess_utils

* unzscore prediction before computing SSIM;
added multi-threading to tiling

* unzscore predition for 3D inference

* bug fix

* turned off normalization for reading 3D target images for computing metrics

* change output dtype to float32

* bug fix

* made metrics_script backward compatible

* Rename "workers" to "num_workers" in the training config

* fixed tests

* fixed tests

* added pool config file

* made Maskgenerator backward compatible

* Add large 2D inference

* bug fix

* computed metrics stats for single FOV

* update inference_script.py

* bug fix

* generate mask meta for user supplied masks; add watershed

* update config files

* update notebook

* update conda env yaml

* edit config

* edit notebook

* Add README for the notebook

* edit comment

* update comment blocks

* update comment blocks

* update README.md

* update notebook

* update README.md

* fix plots not displayed issue

* update README.md

* adding a shell script for setup

* shell script for setup

* update setup script

* update README.md

* update notebook

* edited the image translation exercies for clarity, added TOC, and added jupyterlab to conda environment config.

* update README

* update the paths to data to avoid conflict with 04_image_translation repo

* update instructions

* move README to course repo, clean up the notebook

* fix typos

* update paths for backup tiles

* bux fix

* update plotting

* add config for 2.5D model; add warning for too small min_fraction

* fix margin issue with 2.5D inference

* fix indexing issues in uniform tiling

* bug fix

* bug fix

* bug fix

* bug fix

* bug fix

* fix plotting bug

* fix tiling z indexing bug

* fix inference 2.5D model bug

* fix inference 2.5D model bug, cleaned

* inference: add progressbar, colorbar for figure-target, decrease margin in figure

* fix figure metric assignment

* fix inference input middle slice selection

* plot multiple inputs in figures

* add tqdm to requirements.txt file

* inference refactoring

* changed pd version, convert nan to none when reading meta

* fixed generate meta tests

* fixed inference script tests

* fixed metrics tests

* fixed preprocessing (masks, tile, mp aux)

* fixed evaluation metrics tests

* lots of issues with inference. fixed 2d so far

* fixed 2.5d inference, still working on 3d

* 3 remaining errors in 3d inference

* debugging 3d inference

* debugging 3d inference

* debugging 3d inference, fixed overlap shape

* fixed 5d tiling bug

* fixed 3d inference!

* fixed stitch tests

* fixed dataset tests

* fixed dataset w mask tests

* fixed inference dataset tests

* fixed plot utils tests

* added flatfield tests

* fixed gen mask tests

* working on tiling tests

* fixed tile nonuni tests

* fixed uniform tile tests

* fixed aux utils tests

* debugging image utils

* updated pandas version to avoid attribute error in pandas

* fixed image utils tests

* fixed mask utils tests

* fixed mp utils tests

* debugging tile utils

* fixed tile utils tests

* newer mpl version

* updated skimage, debugging flatfield

* sort output of os.listdir

* changed pd version, convert nan to none when reading meta

* making tests compatible with progress bar changes, still debugging 3d

* removed requirement to run xy metrics, fixed tests for 3d inference

* fixed plot utils test

* sorted glob output

+++ micro_dl/preprocessing/tile_uniform_images.py
@@ -30,7 +30,8 @@ class ImageTilerUniform:
                  int2str_len=3,
                  normalize_im='stack',
                  min_fraction=None,
-                 tile_3d=False):
+                 tile_3d=False,
+                 tiles_exist=False):
         """
         Tiles images.
         If tile_dir already exist, it will check which channels are already
@@ -55,7 +56,7 @@ class ImageTilerUniform:
             indicating if channel should be normalized or not.
         :param int slice_ids: Index of which focal plane acquisition to
             use (for 2D). default=-1 for the whole z-stack
-        :param int pos_ids: Position (FOV) indices to use
+        :param list/int pos_ids: Position (FOV) indices to use
         :param list hist_clip_limits: lower and upper percentiles used for
             histogram clipping.
         :param str flat_field_dir: Flatfield directory. None if no flatfield
@@ -67,6 +68,8 @@ class ImageTilerUniform:
         :param bool tile_3d: Whether tiling is 3D or 2D
          in file names
         :param None or str normalize_im: normalization scheme for input images
+        :param bool tiles_exist: If tiles from channels/masks exist while tiling weights,
+            don't delete previously tiled channels
         """
         self.input_dir = input_dir
         self.output_dir = output_dir
@@ -90,11 +93,12 @@ class ImageTilerUniform:
             self.str_tile_step,
         )
 
-        self.tiles_exist = False
-        # Delete the old tile dir if it already exists
-        if os.path.exists(self.tile_dir):
-            shutil.rmtree(self.tile_dir)
-        os.makedirs(self.tile_dir)
+        self.tiles_exist = tiles_exist
+        if tiles_exist is False:
+            # Delete the old tile dir if it already exists
+            if os.path.exists(self.tile_dir):
+                shutil.rmtree(self.tile_dir)
+        os.makedirs(self.tile_dir, exist_ok=True)
 
         # make dir for saving individual meta per image, could be used for
         # tracking job success / fail
@@ -156,7 +160,7 @@ class ImageTilerUniform:
             assert len(normalize_channels) == len(self.channel_ids),\
                 "Channel ids {} and normalization list {} mismatch".format(
                     self.channel_ids,
-                    self.normalize_channels,
+                    normalize_channels,
                 )
 
             normalize_channels = [normalize_im if flag else None for flag in normalize_channels]
@@ -164,7 +168,6 @@ class ImageTilerUniform:
             self.normalize_channels = \
                 dict(zip(self.channel_ids, normalize_channels))
                 # If more than one depth is specified, length must match channel ids
-
 
     def get_tile_dir(self):
         """
@@ -396,6 +399,7 @@ class ImageTilerUniform:
         else:
             # Using masks, need to make sure they're bool
             is_mask = True
+
         if task_type == 'crop':
             cur_args = (tuple(input_fnames),
                         flat_field_fname,
